+++ CHANGELOG.md
@@ -12,6 +12,7 @@ Version 0.1.2.0
     instances for most of the data types, for auto-generation and searches.
 *   Add some functions for creating `Append`s and witnesses of concatenation
     type families from `Append`s.
+*   `interleavedIxes`, for more manipulation of `Interleave`
 
 Version 0.1.1.1
 ---------------